Refactor ISPyB integration: replace SOAP client with PyISPyB REST client
Replace the legacy SOAP-based ISPyB client with a new PyISPyB REST implementation
backed by Keycloak token authentication. Introduce ``PyISPyBDataAdapter`` and
``PyISPyBRestClient``as abstract, reusable components, removing the old ``ISPyBAdapter``.
Add adjustable timeouts, improved error handling, and legacy method compatibility.
Update tests to reflect the new REST implementation and handle Max IV specific  ``duo``
and ``sdm`` dependencies gracefully.
