+++ mxcubecore/HardwareObjects/MAXIV/ISPyBLims.py
@@ -1,79 +1,15 @@
 # ruff: noqa: TD003, FIX002, ERA001
-import logging
-
-try:
-    # duo is a part of sdm package that lives at MaxIV private repository
-    from duo.UO import RestDuo
-    from sdm.config import DUOPASSWORD, DUOUSER
-except ImportError:
-    RestDuo = None
-    DUOPASSWORD = None
-    DUOUSER = None
-
-from mxcubecore.HardwareObjects.abstract.PyISPyBDataAdapter import PyISPyBDataAdapter
 from mxcubecore.HardwareObjects.abstract.PyISPyBRestClient import PyISPyBRestClient
+from mxcubecore.HardwareObjects.MAXIV.MAXIVPyISPYyBDataAdapter import (
+    LAZY_SESSION_PREFIX,
+    MAXIVPyISPyBDataAdapter,
+)
 from mxcubecore.HardwareObjects.UserTypeISPyBLims import UserTypeISPyBLims
-from mxcubecore.model.lims_session import Proposal, Session
-
-DUO_API_URL = "https://duo-api.maxiv.lu.se"
-LAZY_SESSION_PREFIX = "lazy"
-
-log = logging.getLogger("ispyb_client")
+from mxcubecore.model.lims_session import Session
 
 
 class NoSessionException(Exception):
     """Exception raised when no expected session found."""
-
-
-class CustomISPyBDataAdapter(PyISPyBDataAdapter):
-    """Extend the standard ISPyB data adapter with MAXIV specific logic."""
-
-    def get_proposals(self):
-        """Override method to filter proposals by the type, state and beamline name.
-
-        Include proposals: of type ``MX`` or ``MB`` in ``Open`` state and assigned
-        to the current beamline. The last is checked via DUO API.
-        """
-        duo = RestDuo(DUO_API_URL)
-        duo.login(DUOUSER, DUOPASSWORD)
-        beamline_proposals_ids = set(duo.get_beamline_proposals(self.beamline_name))
-        return [
-            proposal
-            for proposal in super().get_proposals()
-            if proposal.code in ["MX", "MB"]
-            and proposal.state == "Open"
-            and int(proposal.number) in beamline_proposals_ids
-        ]
-
-    def create_session(self, proposal: Proposal) -> Session:
-        """Create a new Session object for the given proposal and beamline.
-
-        This is a lazy session creation, done automatically on the fly in case
-        no appropriate session is found for the user, proposal and current day.
-        This session is labelled with ``lazy`` prefix and is not posted to
-        Py-ISPYB service until it is selected.
-
-        Args:
-            proposal: Proposal object to create session for
-
-        Returns:
-            Session: Created session object
-        """
-        return Session(
-            code=proposal.code,
-            number=proposal.number,
-            proposal_name=proposal.name,
-            proposal_id=proposal.proposal_id,
-            session_id=f"{LAZY_SESSION_PREFIX}{proposal.proposal_id}",
-            beamline_name=self.beamline_name,
-            title=proposal.title,
-            # At MAXIV we don't care if a session is scheduled, set True as default
-            is_scheduled_time=True,
-            is_scheduled_beamline=True,
-            # TODO@dominikatrojanowska: check if we should set start and end time for the session created on fly, and if so, what time should be set. For now, we just set empty string, and let ISPyB handle it.
-            # "startDate": start_time.strftime("%Y-%m-%dT%H:%M:%S.%fZ"),
-            # "endDate": end_time.strftime("%Y-%m-%dT%H:%M:%S.%fZ"),
-        )
 
 
 class ISPyBLims(UserTypeISPyBLims):
@@ -82,8 +18,8 @@ class ISPyBLims(UserTypeISPyBLims):
         self._rest_client = PyISPyBRestClient(pyispyb_rest_root)
         super().init()
 
-    def _create_data_adapter(self) -> CustomISPyBDataAdapter:
-        return CustomISPyBDataAdapter(
+    def _create_data_adapter(self) -> MAXIVPyISPyBDataAdapter:
+        return MAXIVPyISPyBDataAdapter(
             self._rest_client,
             self.beamline_name,
         )
