+++ mxcubecore/model/lims_session.py
@@ -34,7 +34,6 @@ class Proposal(BaseModel):
     code: str = ""
     number: str = ""
     title: str = ""
-    name: str = ""
     state: str = ""
